refactor: fleamarket 파일 분리
